requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
pymysql
python-dotenv
streamlit
matplotlib
cryptography
yfinance
pymongo
feedparser
requests
beautifulsoup4
lxml
plotly
